pyproject.toml
[project]
name = "research-trend-tracker"
version = "3.0.0"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "anthropic>=0.86.0",
    "asyncpg>=0.31.0",
    "celery[redis]>=5.3",
    "fastapi>=0.135.1",
    "fastmcp>=3.1.0",
    "greenlet>=3.3.2",
    "httpx>=0.28.1",
    "langchain>=1.2.12",
    "langgraph>=1.1.5,<1.2.0",
    "langchain-ollama>=1.0.1",
    "networkx>=3.6.1",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "plotly>=6.6.0",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.24",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "pyvis>=0.3.2",
    "redis>=7.3.0",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.55.0",
    "structlog>=25.5.0",
    "uvicorn>=0.41.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]

[tool.mypy]
files = ["app"]

[[tool.mypy.overrides]]
module = ["plotly.*", "pyvis.*"]
ignore_missing_imports = true

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pandas-stubs>=3.0.0.260204",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.7",
    "testcontainers[postgres,redis]>=4.14.1",
    "types-python-jose>=3.5.0.20250531",
    "types-requests>=2.32.4.20260107",
]
